+++ src/main/java/net/pms/dlna/DLNAImageResElement.java
@@ -25,7 +25,7 @@ import java.util.List;
 import net.pms.dlna.DLNAImageProfile.HypotheticalResult;
 import net.pms.image.ImageFormat;
 import net.pms.image.ImageInfo;
-import net.pms.network.mediaserver.UPNPControl.Renderer;
+import net.pms.network.mediaserver.Renderer;
 
 /**
  * This class is used to represent a {@code <res>} element representing an image
